+++ rushub/src/DcConfig.h
@@ -58,8 +58,8 @@ public:
 	int mUsersLimit;	    ///< User's limit
 	int mCompressionType; ///< Compression type
 
-	unsigned long mMaxNmdcCommandLength;
-	unsigned long mMaxWebCommandLength; ///< Max size of web cmd
+	unsigned mMaxNmdcCommandLength;
+	unsigned mMaxWebCommandLength; ///< Max size of web cmd
 	unsigned mMaxCmdLen[NMDC_TYPE_UNKNOWN + 1];
 	unsigned mMaxNickLen; ///< Max length of nick
 	unsigned mMinNickLen; ///< Min length of nick
